backend/django_app/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
mysqlclient==2.2.4
Pillow==10.1.0
celery==5.3.4
redis==5.0.1
django-redis==5.4.0
boto3==1.29.0
django-storages==1.14.2
stripe==7.0.0
reportlab==4.0.7
django-filter==23.3
drf-spectacular==0.26.5
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2
